Added the ability to load saved copies of the series page and spreadsheets so that the package can be used without the internet. Changed how the SDDS series is handled to reflect changes to the website.

+++ R/readSeriesPage.R
@@ -69,22 +69,30 @@ getSeriesPageURL <- function(series){
 ##' @keywords internal
 ##' 
 ##
-readSeriesPage <- function(series, quiet = TRUE){
+readSeriesPage <- function(series, quiet = TRUE, destDir, deleteFiles){
     
     url <- getSeriesPageURL(series)
-    
-    tmp <- tempfile(pattern = "RBNZ", fileext = ".html")
-    on.exit(unlink(tmp))
-    
-    Foo <- try(download.file(url = url,
-                             destfile = tmp,
-                             mode = 'wb',
-                             quiet = quiet
-                             )
-               )
-    
-    if (inherits(Foo, 'try-error') || Foo > 0){
-        stop('Could not download ', url)
+
+    if (deleteFiles | !file.exists(file.path(destDir, paste0(series, '.html'))) ){
+        
+        tmp <- tempfile(pattern = "RBNZ", fileext = ".html")
+        on.exit(unlink(tmp))
+        
+        Foo <- try(download.file(url = url,
+                                 destfile = tmp,
+                                 mode = 'wb',
+                                 quiet = quiet
+                                 )
+                   )
+        
+        if (inherits(Foo, 'try-error') || Foo > 0){
+            stop('Could not download ', url)
+        }
+
+    } else {
+
+        tmp <- file.path(destDir, paste0(series, '.html'))
+
     }
     
     out <- read_html(tmp)
